User access is now broken into its own class.
